Added redirection and made authorized reqs from frontend

+++ frontend/src/components/ItemCard.jsx
@@ -1,5 +1,6 @@
 /* eslint-disable react/prop-types */
 import { marked } from "marked";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 import ItemCardModal from "./ItemCardModal";
 import { useState } from "react";
@@ -7,12 +8,21 @@ import { useState } from "react";
 const ItemCard = ({ item }) => {
   const [isDeleted, setIsDeleted] = useState(false);
   const renderedMarkdown = marked(item.content.substring(0, 200));
+  const { user } = useAuthContext();
   const handleDelete = async () => {
+    if(!user){
+      console.log("You must be logged in to delete an item")
+      return;
+    } 
     try {
       const response = await fetch(
         `http://localhost:4000/api/items/${item._id}`,
         {
           method: "DELETE",
+          headers:{
+            'Authorization':`Bearer ${user.token}`
+
+          }
         }
       );
       if (response.ok) {
